backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["bion-slmn <bionsol25@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
django-redis = "5.4.0"
djangorestframework = "3.14.0"
psycopg2-binary = "2.9.9"
python-dotenv = "0.21.1"
requests = "2.22.0"
pyjwt = "2.3.0"
cryptography = "3.4.7"
djangorestframework-simplejwt = "5.3.0"
redis = "5.0.3"
gunicorn = "22.0.0"
dj-database-url = "2.2.0"
django-cors-headers = "4.3.1"
psycopg2 = "^2.9.10"
setuptools = "^75.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
